+++ test/fixture/variables.tf
@@ -1,6 +1,15 @@
+# For testing purposes, do not delete
 variable "location" {
-  default = "westeurope"
+  type    = string
+  default = "eastus"
 }
 
-# variable "client_id" {}
-# variable "client_secret" {}
+variable "client_id" {
+  type    = string
+  default = null
+}
+
+variable "client_secret" {
+  type    = string
+  default = null
+}